# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EZ Remote Agent",
    "description": "Installs, signs in and runs coding agents.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/agents/claude/settings": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Claude Code settings",
        "operationId": "getClaudeSettings",
        "responses": {
          "200": {
            "description": "Settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaudeSettingsBody"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "agents"
        ],
        "summary": "Change Claude Code settings",
        "description": "Saves the settings, never downgrading the installed version and restarting Remote Control when its settings change.",
        "operationId": "updateClaudeSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaudeSettingsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaudeSettingsBody"
                }
              }
            }
          },
          "400": {
            "description": "A setting Claude cannot take",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/{agent}/login": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Start signing in",
        "description": "Starts the agent's sign-in, replacing any unfinished one, and returns the link to open.",
        "operationId": "startAgentLogin",
        "parameters": [
          {
            "name": "agent",
            "in": "path",
            "description": "The agent",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Agent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sign-in started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginPrompt"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The agent is not installed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "The agent did not show a sign-in link",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent}/login/code": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Finish Claude sign-in",
        "description": "Passes the code Claude shows after website sign-in to the waiting sign-in.",
        "operationId": "submitAgentLoginCode",
        "parameters": [
          {
            "name": "agent",
            "in": "path",
            "description": "The agent, only claude takes a code",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Agent"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Signed in"
          },
          "400": {
            "description": "The agent does not take a code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "No sign-in is in progress",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "The sign-in did not finish",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent}/logout": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Sign the agent out",
        "description": "Signs the agent's CLI out of its account.",
        "operationId": "logOutAgent",
        "parameters": [
          {
            "name": "agent",
            "in": "path",
            "description": "The agent",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Agent"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Signed out"
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The agent is not installed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "The CLI could not sign out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Agent": {
        "type": "string",
        "description": "A command-line coding agent the manager can install and run.",
        "enum": [
          "claude",
          "codex"
        ]
      },
...
      "ClaudeSettingsBody": {
        "type": "object",
        "description": "How the manager installs, updates and serves Claude Code.",
        "required": [
          "release_channel",
          "remote_control"
        ],
        "properties": {
          "release_channel": {
            "$ref": "#/components/schemas/ReleaseChannel"
          },
          "remote_control": {
            "$ref": "#/components/schemas/RemoteControlSettings"
          }
        }
      },
      "CodeBody": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Why a request failed.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "LoginPrompt": {
        "type": "object",
        "description": "What the person signing in needs: a page to open, and sometimes a code to enter there.",
        "required": [
          "url"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "The one-time code to enter on that page, when the sign-in uses one."
          },
          "url": {
            "type": "string",
            "description": "The sign-in page to open."
          }
        }
      },
...
      "ReleaseChannel": {
        "type": "string",
        "description": "Which Claude Code releases to install: `latest` gets every release as soon as it ships,\n`stable` is about a week behind and skips releases with major regressions.",
        "enum": [
          "latest",
          "stable"
        ]
      },
...
      "RemoteControlSettings": {
        "type": "object",
        "description": "How Claude Code's Remote Control servers run.",
        "required": [
          "enabled",
          "permission_mode",
          "capacity",
          "serve_repositories"
        ],
        "properties": {
          "capacity": {
            "type": "integer",
            "format": "int32",
            "description": "The most sessions each server runs at once, from 1 to 32.",
            "default": 4,
            "maximum": 32,
            "minimum": 1
          },
          "enabled": {
            "type": "boolean",
            "description": "Serve /projects, and the folders chosen, to the Claude app while Claude Code is signed in.",
            "default": true
          },
          "permission_mode": {
            "type": "string",
            "description": "The permission mode for sessions started from the Claude app, such as `auto`.",
            "default": "auto"
          },
          "serve_repositories": {
            "type": "boolean",
            "description": "Serve the repositories in /projects that have no choice of their own.",
            "default": true
          }
        }
      },
...
    }
  }
}